Begin Copy displayer for student
